Adds cursor lock to GvrEditorEmulator
Adds cursor lock when changing yaw pitch and roll in the editor. It makes testing easier as the cursor does not interfere with the rest of the editor when the mouse is clicked to simulate a touch event. The cursor is released when alt or ctrl key are released so the touch functionality on the rest of the screen is not compromised